updated
updated userWater, added updateUser
